(RE-6236) Add main WiX project file
Add a WiX project file that will be the basis of all MSI built with
vanagon. It contains all of the generic logic required to build a MSI,
but will not function on it's own. Each project is required to implement
certain files to fill in the rest of the functionality.